backend/foodgram/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.1.0
django-filter==2.4.0
drf-extra-fields==3.0.2
psycopg2-binary==2.9.5
Pillow==9.3.0
gunicorn==20.1.0
python-dotenv==1.1.0



